add more to assembly example
